+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/JsonExtension.java
@@ -31,7 +31,6 @@ import com.diffplug.spotless.json.gson.GsonStep;
 
 public class JsonExtension extends FormatExtension {
 	private static final int DEFAULT_INDENTATION = 4;
-	private static final String DEFAULT_GSON_VERSION = "2.10.1";
 	private static final String DEFAULT_ZJSONPATCH_VERSION = "0.4.14";
 	static final String NAME = "json";
 
@@ -112,7 +111,7 @@ public class JsonExtension extends FormatExtension {
 			this.indentSpaces = DEFAULT_INDENTATION;
 			this.sortByKeys = false;
 			this.escapeHtml = false;
-			this.version = DEFAULT_GSON_VERSION;
+			this.version = GsonStep.DEFAULT_VERSION;
 			addStep(createStep());
 		}